+++ tests/sentry/incidents/subscription_processor/test_subscription_processor_aci.py
@@ -1,6 +1,5 @@
 import copy
 import math
-from contextlib import contextmanager
 from datetime import timedelta
 from functools import cached_property
 from unittest.mock import MagicMock, call, patch
@@ -13,9 +12,7 @@ from urllib3.response import HTTPResponse
 
 from sentry.constants import ObjectStatus
 from sentry.incidents.subscription_processor import SubscriptionProcessor
-from sentry.incidents.utils.subscription_limits import is_metric_subscription_allowed
 from sentry.incidents.utils.types import QuerySubscriptionUpdate
-from sentry.models.organization import Organization
 from sentry.seer.anomaly_detection.types import (
     AnomalyDetectionSeasonality,
     AnomalyDetectionSensitivity,
@@ -1014,96 +1011,3 @@ class MetricsCrashRateDetectorProcessUpdateTest(ProcessUpdateBaseClass, BaseMetr
                 call("incidents.alert_rules.skipping_update_invalid_aggregation_value"),
             ]
         )
-
-
-class TestIsMetricSubscriptionAllowed:
-    org = MagicMock(spec=Organization)
-
-    @contextmanager
-    def fake_features(self, enabled: set[str]):
-        with patch("sentry.incidents.utils.subscription_limits.features") as mock_features:
-            mock_features.has.side_effect = lambda name, *a, **kw: name in enabled
-            yield
-
-    # -- :incidents is the universal gate --
-
-    def test_no_incidents_disables_events(self) -> None:
-        with self.fake_features(set()):
-            assert is_metric_subscription_allowed(Dataset.Events.value, self.org) is False
-
-    def test_no_incidents_disables_transactions(self) -> None:
-        with self.fake_features({"organizations:performance-view"}):
-            assert is_metric_subscription_allowed(Dataset.Transactions.value, self.org) is False
-
-    def test_no_incidents_disables_eap(self) -> None:
-        with self.fake_features({"organizations:visibility-explore-view"}):
-            assert (
-                is_metric_subscription_allowed(Dataset.EventsAnalyticsPlatform.value, self.org)
-                is False
-            )
-
-    def test_no_incidents_disables_performance_metrics(self) -> None:
-        with self.fake_features({"organizations:on-demand-metrics-extraction"}):
-            assert (
-                is_metric_subscription_allowed(Dataset.PerformanceMetrics.value, self.org) is False
-            )
-
-    def test_no_incidents_disables_unknown_dataset(self) -> None:
-        with self.fake_features(set()):
-            assert is_metric_subscription_allowed("unknown_dataset", self.org) is False
-
-    # -- Events --
-
-    def test_events_enabled(self) -> None:
-        with self.fake_features({"organizations:incidents"}):
-            assert is_metric_subscription_allowed(Dataset.Events.value, self.org) is True
-
-    # -- Transactions --
-
-    def test_transactions_without_performance_view(self) -> None:
-        with self.fake_features({"organizations:incidents"}):
-            assert is_metric_subscription_allowed(Dataset.Transactions.value, self.org) is False
-
-    def test_transactions_with_both_features(self) -> None:
-        with self.fake_features({"organizations:incidents", "organizations:performance-view"}):
-            assert is_metric_subscription_allowed(Dataset.Transactions.value, self.org) is True
-
-    # -- EAP --
-
-    def test_eap_without_explore_view(self) -> None:
-        with self.fake_features({"organizations:incidents"}):
-            assert (
-                is_metric_subscription_allowed(Dataset.EventsAnalyticsPlatform.value, self.org)
-                is False
-            )
-
-    def test_eap_with_both_features(self) -> None:
-        with self.fake_features(
-            {"organizations:incidents", "organizations:visibility-explore-view"}
-        ):
-            assert (
-                is_metric_subscription_allowed(Dataset.EventsAnalyticsPlatform.value, self.org)
-                is True
-            )
-
-    # -- PerformanceMetrics --
-
-    def test_performance_metrics_without_on_demand(self) -> None:
-        with self.fake_features({"organizations:incidents"}):
-            assert (
-                is_metric_subscription_allowed(Dataset.PerformanceMetrics.value, self.org) is False
-            )
-
-    def test_performance_metrics_with_both_features(self) -> None:
-        with self.fake_features(
-            {"organizations:incidents", "organizations:on-demand-metrics-extraction"}
-        ):
-            assert (
-                is_metric_subscription_allowed(Dataset.PerformanceMetrics.value, self.org) is True
-            )
-
-    # -- Unknown / other datasets --
-
-    def test_unknown_dataset_with_incidents(self) -> None:
-        with self.fake_features({"organizations:incidents"}):
-            assert is_metric_subscription_allowed("unknown_dataset", self.org) is True
